update: only update cat by owner

+++ src/cat/cat.service.ts
@@ -4,26 +4,23 @@ import {
   Inject,
   Injectable,
   Scope,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { EnumCatGender } from 'src/common/enum/cat.gender';
 import { AppError } from 'src/common/exceptions/app-error';
-import { DeleteResult } from 'typeorm';
 import { CatDto } from './dto/create-cat';
 import { Cat } from './cat.entity';
 import { CatRepository } from './cat.repository';
-import { REQUEST } from '@nestjs/core';
-import { Request } from 'express';
 import { Account } from 'src/auth/account.entity';
-import { AuthService } from 'src/auth/auth.service';
+import { FoodService } from 'src/food/food.service';
+import { Food } from 'src/food/food.entity';
 
 @Injectable({ scope: Scope.REQUEST })
 export class CatsService {
   constructor(
     @InjectRepository(Cat) private readonly catRepo: CatRepository,
-    @Inject(REQUEST)
-    private readonly request: Request,
-    private readonly authService: AuthService,
+    private readonly foodService: FoodService,
   ) {}
 
   async findAll() {
@@ -36,46 +33,54 @@ export class CatsService {
     }
     return await this.catRepo.findOne({
       where: { id: +id },
-      relations: ['owner'],
+      relations: ['owner', 'foods'],
     });
   }
 
-  async create(catDto: CatDto): Promise<Cat> {
+  async create(catDto: CatDto, by: Account): Promise<Cat> {
     if (
       catDto.gender &&
       !Object.values(EnumCatGender).includes(catDto.gender)
     ) {
-      throw new AppError('Invalid gender');
+      throw new AppError('Invalid gender', 400);
     }
 
-    const currentUser = this.request.user as Account;
-
-    if (currentUser) {
-      const owner = await this.authService.findOneById(currentUser.id);
-      if (owner) {
-        const newCat = await this.catRepo.create({ ...catDto, owner });
-        return await this.catRepo.save(newCat);
-      }
+    if (by) {
+      const newCat = await this.catRepo.create({ ...catDto, owner: by });
+      return await this.catRepo.save(newCat);
     }
-    throw new AppError('Invalid user');
+    throw new UnauthorizedException();
   }
 
-  async delete(id: string): Promise<DeleteResult> {
+  async delete(id: string, by: Account): Promise<any> {
     if (isNaN(+id)) {
       throw new HttpException('Invalid id', HttpStatus.BAD_REQUEST);
     }
-    return this.catRepo.delete({ id: +id });
+    const currentCat = await this.findOne(+id);
+
+    if (!currentCat) {
+      return new AppError('NO CAT FOUND', 404);
+    }
+
+    const isOwner = currentCat.owner.id === by.id;
+    if (isOwner) {
+      return await this.catRepo.delete({ id: +id });
+    }
+    throw new UnauthorizedException();
   }
 
-  async update(cat: Cat): Promise<Cat> {
+  async update(cat: Cat, by: Account): Promise<Cat> {
     const currentCat = await this.findOne(cat.id);
     if (!currentCat) {
-      throw new AppError('NOT FOUND CAT');
+      throw new AppError('NOT CAT FOUND', 404);
     }
-    if (cat.gender && !Object.values(EnumCatGender).includes(cat.gender)) {
-      throw new AppError('Invalid gender');
+    if (by && currentCat.owner.id === by.id) {
+      if (cat.gender && !Object.values(EnumCatGender).includes(cat.gender)) {
+        throw new AppError('Invalid gender', 400);
+      }
+      return await this.catRepo.save({ ...currentCat, ...cat });
     }
-    return await this.catRepo.save({ ...currentCat, ...cat });
+    throw new UnauthorizedException();
   }
 
   async findByAccount(id: number) {
@@ -84,5 +89,24 @@ export class CatsService {
       .where('c.ownerId = :id', { id })
       .getManyAndCount();
     return cats;
+  }
+
+  async feed(id: number, foodIds: number[], by: Account) {
+    const currentCat = await this.findOne(id);
+    if (currentCat.owner.id === by.id) {
+      const foodsByIds = await this.foodService.findByIds(foodIds);
+
+      const legal =
+        foodsByIds.filter((f: Food) => f.minAge > currentCat.age).length === 0;
+
+      if (legal) {
+        currentCat.foods = foodsByIds;
+        await this.catRepo.save(currentCat);
+        return true;
+      }
+      throw new AppError('There are some foods not suit the cat', 400);
+    }
+
+    throw new UnauthorizedException();
   }
 }
